+++ src/firebase/init.js
@@ -18,5 +18,4 @@ const auth = getAuth(app);
 const db = getFirestore(app);
 const provider = new GoogleAuthProvider(app);
 
-export {auth,app,db,provider,createUserWithEmailAndPassword,signInWithEmailAndPassword,onAuthStateChanged,signInWithRedirect,
-    getRedirectResult,signOut}
+export {auth,app,db,provider,createUserWithEmailAndPassword,signInWithEmailAndPassword,onAuthStateChanged,signInWithRedirect,getRedirectResult,signOut}
