--- src/signout.js
@@ -1,4 +1,0 @@
-import { signOut } from "./firebase/auth";
-//VISTA HOME
-// Log out
-
